refactor: extract duplicate code to trait
